NameResolver-Flask/requirements.txt
flask==1.1.1
flask-wtf==0.14.2
flask-sqlalchemy==2.4.1
pymongo==2.9.0
flask-pymongo==2.3.0
Flask-MongoAlchemy==0.7.2
flask-bcrypt==0.7.1
flask-login==0.4.1
pillow==6.2.1
